Task 6: Create tests for the library task: Addition

+++ Calculator-ClassLibrary/Calculator.cs
@@ -5,19 +5,19 @@ namespace Calculator_ClassLibrary
     public static class Calculator
     {
         // Method for adding 2 integers
-        public static double Add(int lValue, int rValue)
+        public static int Add(int lValue, int rValue)
         {
             return lValue + rValue;
         }
 
         // Method for subtracting 2 integers
-        public static double Subtract(int lValue, int rValue)
+        public static int Subtract(int lValue, int rValue)
         {
             return lValue - rValue;
         }
 
         // Method for multiplying 2 integers
-        public static double Multiply(int lValue, int rValue)
+        public static int Multiply(int lValue, int rValue)
         {
             return lValue * rValue;
         }
